refactor: inventories 객체 생성 책임 InventoriesGenerator로 분리
