--- 4-period/java-vendingmachine-precourse-main/src/main/java/vendingmachine/dto/request/InventoriesRequest.java
@@ -1,8 +1,0 @@
-package vendingmachine.dto.request;
-
-import java.util.List;
-
-public record InventoriesRequest(
-        List<InventoryRequest> InventoriesRequest
-) {
-}
