--- 4-period/java-vendingmachine-precourse-main/src/main/java/vendingmachine/dto/request/InventoryRequest.java
@@ -1,8 +1,0 @@
-package vendingmachine.dto.request;
-
-public record InventoryRequest(
-        String name,
-        int price,
-        int stock
-) {
-}
